task:add new vehicle sep comp in progress
