+++ src/components/addNewVehicleToService/CustomerComplaintsInput.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const CustomerComplaintsInput = () => {
+  return <div>CustomerComplaintsInput</div>;
+};
+
+export default CustomerComplaintsInput;
