+++ src/components/addNewVehicleToService/CustomerInformationInput.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+const CustomerInformationInput = () => {
+  return <div>CustomerInformationInput</div>;
+};
+
+export default CustomerInformationInput;
